refactor: move shared changelog utilities
- Move functions shared by multiple changelog scripts to util/changelog.py from the individual scripts
- Reset changeset id for changelog tests
