+++ python/scripts/make_model_changelog.py
@@ -2,10 +2,9 @@
 Script takes one MDF file representing a model and produces a Liquibase 
 Changelog with the necessary changes to add the model to an MDB
 """
-import configparser
 import logging
 from pathlib import Path
-from typing import Dict, Generator, List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import click
 from bento_mdf.mdf import MDF
@@ -13,39 +12,12 @@ from bento_meta.entity import Entity
 from bento_meta.mdb.mdb import make_nanoid
 from bento_meta.model import Model
 from bento_meta.objects import Concept, Term, ValueSet
+from bento_meta.util.changelog import changeset_id_generator, update_config_changeset_id
 from bento_meta.util.cypher.clauses import Create, Match, Merge, OnCreateSet, Statement
 from bento_meta.util.cypher.entities import N, R, T, _plain_var
 from liquichange.changelog import Changelog, Changeset, CypherChange
 
 logger = logging.getLogger(__name__)
-
-
-def get_initial_changeset_id(config_file_path: str) -> int:
-    """Gets initial changeset id from changelog config file"""
-    config = configparser.ConfigParser()
-    config.read(config_file_path)
-    try:
-        return config.getint(section="changelog", option="changeset_id")
-    except (configparser.NoSectionError, configparser.NoOptionError) as error:
-        logger.error(f"Reading changeset ID failed: {error}")
-        raise
-
-
-def changeset_id_generator(config_file_path: str) -> Generator[int, None, None]:
-    """Generates sequential changeset IDs by reading the latest ID from a config file."""
-    i = get_initial_changeset_id(config_file_path)
-    while True:
-        yield i
-        i += 1
-
-
-def update_config_changeset_id(config_file_path: str, new_changeset_id: int) -> None:
-    """Updates changelog config file with new changeset id."""
-    config = configparser.ConfigParser()
-    config.read(config_file_path)
-    config.set(section="changelog", option="changeset_id", value=str(new_changeset_id))
-    with open(file=config_file_path, mode="w", encoding="UTF-8") as config_file:
-        config.write(fp=config_file)
 
 
 def escape_quotes_in_attr(entity: Entity) -> None:
