+++ python/src/bento_meta/objects.py
@@ -91,6 +91,7 @@ class Property(Entity):
             "units": "units",
             "item_domain": "item_domain",
             "is_required": "is_required",
+            "parent_handle": "parent_handle",
         },
         "relationship": {
             "concept": {"rel": ":has_concept>", "end_cls": "Concept"},
