+++ python/src/bento_meta/util/__init__.py
@@ -3,6 +3,7 @@ bento_meta.util - Utilities
 
 bento_meta.util.cypher - Represent Cypher queries programmatically
 bento_meta.util.makeq - Create Cypher queries from API endpoint paths
+bento_meta.util.changelog - Utilities for working with Liquibase Changelogs
 """
 
-from . import cypher
+from . import changelog, cypher
