+++ python/tests/samples/test_changelog.ini
@@ -1,3 +1,3 @@
 [changelog]
-changeset_id = 6855
+changeset_id = 1
